refactor: split App into components
